+++ CHANGELOG.md
@@ -4,6 +4,68 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## 2023-09-31
+## [1.0.0] - 2023-09-31
 ### Added
 - Initial version
+
+## [1.1.0] - 2024-01-10
+### Added
+
+- Use CloudFront KeyValueStore to store commitId
+
+### Changed
+### Fixed
+
+## [1.1.1] - 2024-01-10
+### Added
+
+### Changed
+
+- remove age and date header
+
+### Fixed
+
+## [1.1.2] - 2024-01-10
+### Added
+
+### Changed
+
+- build AWS Layer manually
+
+### Fixed
+
+## [1.1.3] - 2024-01-25
+### Added
+
+- max name for OAC & remove old comments
+
+### Changed
+### Fixed
+
+## [1.1.4] - 2024-01-25
+### Added
+
+### Changed
+
+- use custom name for KVS
+
+### Fixed
+
+## [1.1.5] - 2024-01-25
+### Added
+
+- Use Lambda Power Tools for logging and tracing
+
+### Changed
+### Fixed
+
+
+## [1.1.6] - 2024-02-13
+### Added
+
+- Use Lambda Power Tools for logging and tracing
+
+### Changed
+
+- Use typescript for Lambda function instead of javascript
+### Fixed